build-mcp-server-python-tutorial/pyproject.toml
[project]
name = "github-mcp-server"
version = "1.0.0"
description = "MCP server for comprehensive GitHub API integration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["mcp", "github", "claude", "ai", "automation"]

dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
github-mcp = "github_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/github_mcp"]
